BookStoreAPI/requirements.txt
pytest==8.1.1
requests==2.31.0
pytest-html==4.1.1
pytest-xdist==3.5.0
python-dotenv==0.13.0